Build libgnome-keyring again
It is built and installed as a part of the electron base app, but the
cleanup rules there remove all the pkg-config files and headers. We
require them when building the keytar module in atom.

Maybe the base app should not cleanup anything?

The libgnome-keyring snippet and the patch are stolen from the
electron base app flatpak manifest.